Create Loops
In programming, loops are structures that allow you to execute a block of code repeatedly. Instead of writing the same line of code ten times, you use a loop to tell the computer: "Do this task until a specific condition is met."

In Python, there are two primary types of loops: for loops and while loops.